Refactor modal item editing views
They were all basically the same, so I put them in one file
