--- editorsnotes_app/src/views/add_topic.js
@@ -1,17 +1,0 @@
-"use strict";
-
-var Cocktail = require('backbone.cocktail')
-  , TopicView = require('./topic')
-  , AddItemMixin = require('./generic/add_item_base')
-
-module.exports = Cocktail.mixin(TopicView.extend({}), AddItemMixin, {
-  itemType: 'topic',
-  initialize: function () {
-    this.$('.save-row').remove();
-    this.renderModal();
-  },
-  saveItem: function (e) {
-    e.preventDefault();
-    this.model.save().then(this.$el.modal.bind(this.$el, 'hide'));
-  }
-});
